+++ src/app/app.module.ts
@@ -13,6 +13,7 @@ import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
 import { ProfileModule } from './modules/profile/profile.module';
 import { AuthService } from './services/auth.service';
+import {CookieEncryptionHelper} from "./helpers/cookie-encryption.helper";
 
 
 export function HttpLoaderFactory(http: HttpClient) {
@@ -24,7 +25,6 @@ export function HttpLoaderFactory(http: HttpClient) {
 @NgModule({
   declarations: [
     AppComponent,
-   
   ],
   imports: [
     BrowserModule,
@@ -47,7 +47,8 @@ export function HttpLoaderFactory(http: HttpClient) {
     })
   ],
   providers: [
-    AuthService
+    AuthService,
+    CookieEncryptionHelper
   ],
   bootstrap: [AppComponent]
 })