+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). 
 
+### [ita-challenges-frontend-2.0.62-RELEASE] - 2024-03-13
+* Add helpers and encryption cookies.
+
 ### [ita-challenges-frontend-2.0.60-RELEASE] - 2024-03-07
 * Change rename environment and delete dummy variables name"